+++ src/app/layout.js
@@ -1,4 +1,6 @@
 import { Inter } from "next/font/google";
+import { SongsContextProvider } from "./context-hook/useSongsState";
+
 import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
@@ -11,7 +13,9 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <SongsContextProvider>{children}</SongsContextProvider>
+      </body>
     </html>
   );
 }
