+++ src/app/page.js
@@ -2,8 +2,9 @@
 
 import Image from "next/image";
 import styles from "./page.module.css";
-import { useState, useEffect } from "react";
-import axios from "axios";
+// import { useState, useEffect } from "react";
+// import axios from "axios";
+import { songs } from "./context-hook/useSongsState";
 import SongsForm from "./components/songsForm";
 import Link from "next/link";
 import Composition from "./components/composition";
@@ -12,20 +13,20 @@ export default function Home() {
   const [openModal, setOpenModal] = useState(false);
 
   //up to date list of songs
-  const [songs, setSongs] = useState([]);
+  // const [songs, setSongs] = useState([]);
 
-  useEffect(() => {
-    fetchSongs();
-  }, []);
+  // useEffect(() => {
+  //   fetchSongs();
+  // }, []);
 
-  const fetchSongs = async function () {
-    try {
-      const res = await axios.get("/api/songs");
-      setSongs(res.data);
-    } catch (error) {
-      console.error("error fetching songs:", error);
-    }
-  };
+  // const fetchSongs = async function () {
+  //   try {
+  //     const res = await axios.get("/api/songs");
+  //     setSongs(res.data);
+  //   } catch (error) {
+  //     console.error("error fetching songs:", error);
+  //   }
+  // };
 
   return (
     <main className={styles.main}>
@@ -45,7 +46,6 @@ export default function Home() {
         close={() => {
           setOpenModal(false);
         }}
-        updateSongs={fetchSongs}
       />
 
       <div className={styles.grid}>
